YEAHSSS YEAHSSS IT WORKSSSS i uhhh firmware and app now works in the sense that: i fixed an untested firmware error where there was an uninitialized pointer being dereferenced before setup was even called, i thank to the more advanced ai chatbot that found that out instead of trying to convince me my setup function is wrong also forgot to make the Reply object so fixed that too then, i am happy with the app, made a logging system that i am very happy about now that i think about it, i think i will just use this one app for everything and make it better and better :3 do not know how multiple devices will get handled or how will devices handle multiple masters so to end it haven't tested thouroughly much, but i can get the app to locate the device with broadcast pings
time to sleep.
i am happyyy :33
also hello favo i luv u :333 (easter egg)
(hope this time it will commit correctly)

+++ src/ESP12E/src/commands.h
@@ -1,6 +1,6 @@
-#define COMMAND_RESV 0
+#define COMMAND_RESV (uint8_t)0
 #define COMMAND_PING (uint8_t)1
-#define COMMAND_RGBV 2
-#define COMMAND_CONNRESET 3
-#define COMMAND_READREG 8
-#define COMMAND_WRITEREG 9
+#define COMMAND_RGBV (uint8_t)2
+#define COMMAND_CONNRESET (uint8_t)3
+#define COMMAND_READREG (uint8_t)8
+#define COMMAND_WRITEREG (uint8_t)9
